I've also updated the javax.jar file and replaced the abstract classes with the jdbc2 ones.  and changed the firebird/* build/install scripts to not required the whole j2ee install.
Although the packaging instructions work, they obscure a bit building the interclient in a simple manner.  So I updated the README with some simple instructions to build interserver and the interclient.jar.  (I also added a 4 line Makefile.linux to compile the classes).  To build a package follow Tom's mod's of Borlands instructions, but to play arround you can do these.

I deleted the obfuscatedClasses, packages/coffee, and packages/crema trees, these seem the only be used in the obsure build and are no longer required.

I also applied any changes made to the 1.6 tree to the 2.0 tree.  These are:

1. All of  Torsten's mods mostly depreciation warnings, including some datetime stamp things and a few String(byte[], 0) call changes.

2.  fix for match on label and or name of column ResultSet.findColumn().

3.  New feature Statement.getFetchSize() now returns a result (result was already around).

4.  Test for end of a result set was in wrong place, could cause exceptions in some circumstances ResultSet.getNextCursorPosition().

(Actually I think Torsten was the instigator for most of the above).